fix(core): distinguish palette source failures
Copilot CLI: 1.0.79
Model: GPT-5.6 Sol (gpt-5.6-sol)

+++ FEBuilderGBA.Avalonia/ViewModels/ImageUnitPaletteViewModel.cs
@@ -231,7 +231,7 @@ namespace FEBuilderGBA.Avalonia.ViewModels
         {
             ROM rom = CoreState.ROM;
             if (rom == null || !CanWrite || CurrentAddr == 0 ||
-                CurrentAddr + SIZE > (uint)rom.Data.Length)
+                (ulong)CurrentAddr + SIZE > (ulong)rom.Data.Length)
             {
                 return false;
             }
